first update
First, create a game folder and put css html js files in it. After putting them in the game folder, create a file named img in that game folder. Put .pngs in it.
Click on the html file to run it.
Player 2 moves with the a and d keys,
and attacks with shift.
Player 1 moves left and right with the arrow keys and moves with the space bar.